Rename modules and fix notebook imports (#68)
* rename superop_conversion to superoperator_conversion in python files and notebooks

* change superop_conversion to superoperator_conversion in python files and notebooks

* rpe -> robust_phase_estimation

* rb -> randomized_benchmarking

* dfe -> direct_fidelity_estimation

* bell -> bell_state

* fix imports was forest_benchmarking changed to forest.benchmarking
