+++ forest/benchmarking/tests/test_direct_fidelity_estimation.py
@@ -1,4 +1,4 @@
-from forest.benchmarking.dfe import exhaustive_state_dfe, exhaustive_process_dfe, ratio_variance, \
+from forest.benchmarking.direct_fidelity_estimation import exhaustive_state_dfe, exhaustive_process_dfe, ratio_variance, \
     monte_carlo_process_dfe
 from pyquil import Program
 from pyquil.api import BenchmarkConnection
