+++ forest/benchmarking/tests/test_randomized_benchmarking.py
@@ -3,7 +3,7 @@ from numpy import random, uint8, zeros
 
 from pyquil.gates import X, Y
 from pyquil.quil import Program
-from forest.benchmarking.rb import merge_sequences, fit_standard_rb, rb_dataframe, \
+from forest.benchmarking.randomized_benchmarking import merge_sequences, fit_standard_rb, rb_dataframe, \
     add_sequences_to_dataframe, run_rb_measurement, survivals_by_qubits, add_survivals, survival_statistics, \
     fit_unitarity, add_unitarity_sequences_to_dataframe, \
     run_unitarity_measurement, add_shifted_purities, shifted_purities_by_qubits, unitarity_to_RB_decay
