+++ forest/benchmarking/tomography.py
@@ -11,7 +11,7 @@ from scipy.linalg import logm, pinv, eigh
 
 import forest.benchmarking.distance_measures as dm
 import forest.benchmarking.operator_estimation as est
-from forest.benchmarking.superop_conversion import vec, unvec
+from forest.benchmarking.superoperator_conversion import vec, unvec
 from forest.benchmarking.utils import prepare_prod_sic_state, n_qubit_pauli_basis, partial_trace
 from pyquil import Program
 from pyquil.api import QuantumComputer
